refactor: broke up util file and renamed
